Move values extraction to pipeline_funs.py
